+++ src/js/config.js
@@ -1,5 +1,7 @@
 import mi18n from 'mi18n'
 
+const noop = () => null
+
 // eslint-disable-next-line
 mi18n.addLanguage('en-US', FB_EN_US)
 
@@ -43,11 +45,15 @@ export const defaultOptions = {
     success: console.log,
     warning: console.warn,
   },
-  onAddField: (fieldData, fieldId) => fieldData,
-  onClearAll: () => null,
-  onCloseFieldEdit: () => null,
-  onOpenFieldEdit: () => null,
-  onSave: (evt, formData) => null,
+  onAddField: (fieldId, fieldData) => fieldData,
+  onClearAll: noop,
+  onCloseFieldEdit: noop,
+  onOpenFieldEdit: noop,
+  /**
+   * @param {Object} evt
+   * @param {Object} formData
+   */
+  onSave: noop,
   prepend: false,
   replaceFields: [],
   roles: {