+++ src/js/helpers.js
@@ -743,9 +743,13 @@ export default class Helpers {
    * Toggles the edit mode for the given field
    * @param  {String} fieldId
    * @param  {Boolean} animate
+   * @return {Node|null} field
    */
   toggleEdit(fieldId, animate = true) {
     const field = document.getElementById(fieldId)
+    if (!field) {
+      return field
+    }
     const $editPanel = $('.frm-holder', field)
     const $preview = $('.prev-holder', field)
     field.classList.toggle('editing')
@@ -766,6 +770,7 @@ export default class Helpers {
       config.opts.onCloseFieldEdit($editPanel[0])
       document.dispatchEvent(events.fieldEditClosed)
     }
+    return field
   }
 
   /**
@@ -871,7 +876,9 @@ export default class Helpers {
       return false
     }
 
-    if (!fieldID) {
+    const field = fieldID && document.getElementById(fieldID)
+
+    if (!fieldID || !field) {
       const availableIds = [].slice.call(fields).map(field => {
         return field.id
       })
@@ -881,7 +888,6 @@ export default class Helpers {
       fieldID = form.lastChild.id
     }
 
-    const field = document.getElementById(fieldID)
     const $field = $(field)
     if (!field) {
       config.opts.notify.warning('Field not found')