+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [3.3.1](https://github.com/kevinchappell/formBuilder/compare/v3.3.0...v3.3.1) (2020-01-28)
+
+
+### Bug Fixes
+
+* field removed from stage but not formData ([2c6d824](https://github.com/kevinchappell/formBuilder/commit/2c6d824fe5dcbe413440b3c3fa59b5944e1c5bf9)), closes [#1028](https://github.com/kevinchappell/formBuilder/issues/1028)
+
+# [3.3.0](https://github.com/kevinchappell/formBuilder/compare/v3.2.6...v3.3.0) (2020-01-26)
+
+
+### Bug Fixes
+
+* **demo:** clear current id when removed from stage ([e0c0f2e](https://github.com/kevinchappell/formBuilder/commit/e0c0f2ea8f204da6e11d8de938e27ca6687d5588))
+
+
+### Features
+
+* getCurrentFieldId ([c0148b3](https://github.com/kevinchappell/formBuilder/commit/c0148b33dc7154e2bc2a241255b7a2db6257534f)), closes [#571](https://github.com/kevinchappell/formBuilder/issues/571)
+
+## [3.2.6](https://github.com/kevinchappell/formBuilder/compare/v3.2.5...v3.2.6) (2019-12-31)
+
+
+### Bug Fixes
+
+* adblock because of fb- prefix ([e091474](https://github.com/kevinchappell/formBuilder/commit/e091474219cba88ff0c950649de62ec0de9ab414))
+
 ## [3.2.5](https://github.com/kevinchappell/formBuilder/compare/v3.2.4...v3.2.5) (2019-06-26)